fix(AuthRepository): make userId Observable
Previously, userId would sometimes return undefined, because the Firebase Auth SDK would not have fully initialized yet. With this fix, the observable will emit the current user id as soon as the Auth SDK initializes, so it is no longer possible to get userId of undefined when the user is signed in.