ux: prevent SideBar and UserListItem flickering on page load
These components needed to access the information about whether the user is authenticated in order to render, and thus would hide some of their elements until that data is available. With this fix, page will not be rendered until the auth data is available. Therefore, the ui will not flicker